requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
pyyaml==6.0.2
psutil==6.1.1
python-multipart==0.0.20
WSDiscovery==2.1.0
aiofiles==24.1.0
